More unit tests for events views
