+++ dthm4kaiako/tests/events/urls/test_upcoming.py
@@ -1,8 +1,6 @@
 """Unit tests for upcoming events url"""
 
 from django.urls import reverse, resolve
-from http import HTTPStatus
-from django.test.utils import override_settings
 from tests.BaseTestWithDB import BaseTestWithDB
 
 
@@ -17,9 +15,3 @@ class UpcomingEventsURLTest(BaseTestWithDB):
 
     def test_upcoming_resolve_provides_correct_view_name(self):
         self.assertEqual(resolve("/events/upcoming/").view_name, "events:upcoming")
-
-    @override_settings(GOOGLE_MAPS_API_KEY="mocked")
-    def test_upcoming_gives_200_status_code(self):
-        url = reverse("events:upcoming")
-        response = self.client.get(url)
-        self.assertEqual(HTTPStatus.OK, response.status_code)
