+++ dthm4kaiako/tests/events/views/test_emergency_details_valid.py
@@ -1,0 +1,1 @@
+"""Unit tests for emergency_details_valid"""
